+++ skbio/io/format/fastq.py
@@ -317,7 +317,12 @@ def _fastq_sniffer(fh):
 
     try:
         not_empty = False
-        for _ in zip(range(10), _fastq_to_generator(fh, phred_offset=33)):
+        for _, seq in zip(range(10), _fastq_to_generator(fh, phred_offset=33)):
+            split_length = len((seq.metadata['id'] +
+                                seq.metadata['description']).split(':'))
+            description = seq.metadata['description'].split(':')
+            if split_length == 10 and description[1] in 'YN':
+                return True, {'variant': 'illumina1.8'}
             not_empty = True
         return not_empty, {}
     except (FASTQFormatError, ValueError):